+++ packages/moonstone/Scroller/Scrollable.js
@@ -146,6 +146,7 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 		isFirstDragging = false
 		isDragging = false
 		isKeyDown = false
+		isPrerender = true
 
 		// event handlers
 		eventHandlers = {}
@@ -673,6 +674,8 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 					});
 				}
 			}
+
+			this.isPrerender = false;
 		}
 
 		componentDidUpdate () {
@@ -733,8 +736,8 @@ const ScrollableHoC = hoc((config, Wrapped) => {
 					className
 				),
 				isBothScrollable = isHorizontalScrollbarVisible && isVerticalScrollbarVisible,
-				verticalScrollbarClassnames = isVerticalScrollbarVisible ? (!isBothScrollable && css.onlyVerticalScrollbarNeeded) : css.verticalScrollbarDisabled,
-				horizontalScrollbarClassnames = isHorizontalScrollbarVisible ? (!isBothScrollable && css.onlyHorizontalScrollbarNeeded) : css.horizontalScrollbarDisabled;
+				verticalScrollbarClassnames = isVerticalScrollbarVisible && !this.isPrerender ? (!isBothScrollable && css.onlyVerticalScrollbarNeeded) : css.verticalScrollbarDisabled,
+				horizontalScrollbarClassnames = isHorizontalScrollbarVisible && !this.isPrerender ? (!isBothScrollable && css.onlyHorizontalScrollbarNeeded) : css.horizontalScrollbarDisabled;
 
 			delete props.className;
 			delete props.cbScrollTo;